refactor: apply architecture-review findings — honest module names, typed example registry
- execute-images.ts -> execute.ts: the module owns the whole
  figma_execute_code orchestration, not just the image branch.
- figma-execute-code-tool.ts -> code-mode-tools.ts: it registers the
  three-tool code-mode group, so the name says so instead of reading as
  a break of the one-file-per-tool convention.
- examples.ts metadata now keyed Record<keyof typeof EXAMPLE_CODE, ...>
  so both drift directions fail compile — previously a new example with
  no metadata generated cleanly, passed docs:check, and silently never
  rendered in the reference.
