+++ CLAUDE.md
@@ -8,7 +8,7 @@ This repo is a pnpm workspace. The **root package `figma-developer-mcp`** is the
 
 - **`plugin/`** (`@framelink/plugin`) — the Figma plugin: the `flcm` authoring DSL preamble (`src/preamble/`, an esbuild IIFE bundled into the QuickJS sandbox), `code.ts` host, `manifest.json`, `ui.html`, `build.mjs`. Its schema (`src/preamble/schema.ts`) is the write edge of the canonical vocabulary — the root build bundles it into `dist` via tsup `noExternal` (the package itself is never published).
 
-The write path lives in the root `src/`: `src/services/plugin-bridge/` is the WS relay to the plugin (port block, approval gate, version-skew policy, the trusted image fetch), and `src/mcp/tools/figma-execute-code-tool.ts` registers the code-mode tools (`figma_execute_code`, `get_flcm_reference`, `get_screenshot`). The relay starts with the server in both transports; the tools are **dynamically advertised** — disabled until a plugin connects (latched on for the process lifetime; `--code-mode` forces always-on). `src/mcp/tools/flcm-docs/` generates the tool docs from the schema; `scripts/gen-flcm-doc.ts` regenerates the committed artifacts (`pnpm docs:gen`, drift-gated by `pnpm docs:check`).
+The write path lives in the root `src/`: `src/services/plugin-bridge/` is the WS relay to the plugin (port block, approval gate, version-skew policy, the trusted image fetch), and `src/mcp/tools/code-mode-tools.ts` registers the code-mode tools (`figma_execute_code`, `get_flcm_reference`, `get_screenshot`). The relay starts with the server in both transports; the tools are **dynamically advertised** — disabled until a plugin connects (latched on for the process lifetime; `--code-mode` forces always-on). `src/mcp/tools/flcm-docs/` generates the tool docs from the schema; `scripts/gen-flcm-doc.ts` regenerates the committed artifacts (`pnpm docs:gen`, drift-gated by `pnpm docs:check`).
 
 flcm design docs (ADRs, plans, sketches, solutions) live in **`docs/flcm/`** — note `docs/` is a separate private repo (FramelinkAI/mcp-docs) cloned at that path and gitignored here; design docs must not be committed to this public repo. The tracked generated docs are `plugin/docs/authoring/flcm.md` and `src/mcp/tools/flcm-docs/examples-code.generated.ts` (CI-gated, regenerated from code).
